+++ app/Http/Controllers/Admin/ProductController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Events\ProductCreated;
 use App\Events\ProductUpdated;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\ProductRequest;
@@ -9,9 +10,12 @@ use App\Models\Attribute;
 use App\Models\Brand;
 use App\Models\Category;
 use App\Models\Product;
+use App\Traits\PreventsSourcedResourceDeletion;
 
 class ProductController extends Controller
 {
+    use PreventsSourcedResourceDeletion;
+
     /**
      * Display a listing of the resource.
      *
@@ -115,6 +119,11 @@ class ProductController extends Controller
     public function destroy(Product $product)
     {
         abort_unless(request()->user()->is('admin'), 403, 'You don\'t have permission.');
+
+        if (($result = $this->preventSourcedResourceDeletion($product)) !== true) {
+            return $result;
+        }
+
         $product->delete();
 
         return request()->ajax()
